refactored export data
