+++ flask_monitoringdashboard/templates/dashboard/overview.html
@@ -56,7 +56,7 @@
       <div class="panel-heading">
         <a href="{{ url_for('dashboard.download_csv') }}"
            class="btn btn-primary btn-sm">Download CSV</a>
-        <a href="{{ url_for('dashboard.export_data') }}"
+        <a href="{{ url_for('dashboard.view_csv') }}"
            class="btn btn-primary btn-sm">View CSV</a>
       </div>
     </div>
