+++ flask_monitoringdashboard/test/test_config.py
@@ -1,7 +1,5 @@
 import unittest
 
-from flask_monitoringdashboard.test.utils import get_test_app
-
 
 class TestConfig(unittest.TestCase):
