+++ flask_monitoringdashboard/views/export/json.py
@@ -1,58 +1,12 @@
 import datetime
 
 import jwt
-from flask import make_response, render_template, session, request, json, jsonify
+from flask import json, jsonify
 
 from flask_monitoringdashboard import blueprint, config
-from flask_monitoringdashboard.database.function_calls import get_data, get_data_between
+from flask_monitoringdashboard.database.function_calls import get_data_between
 from flask_monitoringdashboard.database.monitor_rules import get_monitor_data
-from flask_monitoringdashboard.database.tests import add_or_update_test, add_test_result, get_suite_nr
-from flask_monitoringdashboard.database.tests_grouped import reset_tests_grouped, add_tests_grouped
-from flask_monitoringdashboard.core.auth import admin_secure
-# from setup import VERSION
 from flask_monitoringdashboard.views.details.utils import get_details
-
-CSV_COLUMNS = ['endpoint', 'execution_time', 'time', 'version', 'group_by', 'ip']
-
-
-@blueprint.route('/download-csv')
-@admin_secure
-def download_csv():
-    csv = ','.join(CSV_COLUMNS) + '\n'
-    for entry in get_data():
-        csv += ','.join([str(entry.__getattribute__(c)) for c in CSV_COLUMNS]) + '\n'
-
-    response = make_response(csv)
-    response.headers["Content-Disposition"] = "attachment; filename=measurements_{0}.csv".format(
-        str(datetime.datetime.now()).replace(" ", "_").replace(":", "-")[:19])
-    return response
-
-
-@blueprint.route('/export-data')
-@admin_secure
-def export_data():
-    csv = [','.join(CSV_COLUMNS)]
-    data = get_data()
-    for entry in data:
-        csv.append(','.join([str(entry.__getattribute__(c)) for c in CSV_COLUMNS]) + '\n')
-    return render_template('export-data.html', link=config.link, session=session, data=csv)
-
-
-@blueprint.route('/submit-test-results', methods=['POST'])
-def submit_test_results():
-    content = request.get_json()['test_runs']
-    suite = get_suite_nr()
-    for result in content:
-        time = datetime.datetime.strptime(result['time'], '%Y-%m-%d %H:%M:%S.%f')
-        add_or_update_test(result['name'], time, result['successful'])
-        add_test_result(result['name'], result['exec_time'], time, config.version, suite, result['iter'])
-
-    groups = request.get_json()['grouped_tests']
-    if groups:
-        reset_tests_grouped()
-        add_tests_grouped(groups)
-
-    return '', 204
 
 
 @blueprint.route('/get_json_data', defaults={'time_from': 0})
